Split the tests from one monolithic test start and end to per test suite. Fixed some more bugs as usual.

+++ components/utils/tick.h
@@ -21,7 +21,7 @@ void systick_init(void);
 void dwt_cnt_init(void);
 
 // Get number of milliseconds since boot
-uint32_t ms_since_boot(void);
+uint64_t ms_since_boot(void);
 
 
 // Polling delay macros
